Do not auto-renegotiate OTR if not currently encrypting
Fixes an issue where doing a Finish re-negotiates if remote thinks we are not
encrypting. If the remote isn't encrypting, they will return an error. Do not
initiate OTR in this case. Solution: do not auto-negotiate OTR if we get an
error but are not currently encrypting.

fixes #1545 https://dev.guardianproject.info/issues/1545