Comply to spec when reading data message old mac
The "Old MAC keys to be revealed" field is a DATA field according to
the OTR protocol specification of all versions. Hence, reading it as a
(single) mac address of 20 bytes is wrong in two senses:
1. it might be empty or contain more than one address, which would be
   ignored then
2. a DATA field contains a 4 byte length header, which would be read as
   the first 4 bytes of the first mac address.

This commit correctly parses the field as a DATA field.

(cherry picked from commit b05ec27611bc930b8fe1423dba3a6e211967127b)