Use a single SecureRandom instance per Session.
Modified the code to contain only a single instance of SecureRandom.
The instance is stored in Session. A package-private method
'secureRandom()' can be used to access the instance for use by other
parts, as all parts share the same Session instance already.

The SecureRAndom instance is intended to be dedicated to a single session.
It is shared among slave sessions, as these are all related to the same
master session.

The few crypto classes that require SecureRandom will get an instance
passed on as a parameter, to avoid unnecessary instantiation of
SecureRandom.

SecureRandom is immediately created in the constructor of Session, as
it is needed anyways. The instance is saved into an instance variable
and is final, as further modifications are not applicable.