Improved instance tag parsing.
* Extended instance tag parsing with verification of maximum instance
tag length. This check is now required as BigInteger can support
vastly larger numbers. Subsequently calling intValue() will just
truncate anything that does not fit in 32 bits. We want to know
beforehand if the tag fits 32 bits, if not this is already a violation
of the protocol.
* Added tests for valid 32-bit long instance tag and invalid 33-bit long
instance tag.
* Added FIXME comments for other checks that should be implemented in
order to do more complete verification of protocol correctness.