Fix startSession condition when evaluating OTR policy.
The original if-condition stated that if either OTRv2 or OTRv3 is
disabled in the active OTR policy, then we should throw an
UnsupportedOperationException. This does not make sense, since if
either one of the versions is disabled by the policy, we can still fall
back on the other option.
What we should want is that if both OTRv2 and OTRv3 is not an option
because both are disabled, then we have no fallback option and *then* we
throw an UnsupportedOperationException stating that we have no valid
options left.