session: redesign, separate Session and Instance, remove default outgoing session
This change introduces a significant change in the API of otr4j. otr4j
originally was designed with one session in mind. OTRv2 only knew about
a (single) session for an account that interacts with a contact. With the
introduction of OTRv3, instance-tags were introduced to accomodate for
multiple clients being active simultaneously for a single account:
instance-tags would allow distinguishing between multiple clients, all
of which (attempt to) establish an OTR encrypted session at the same
time. This change resolved the constant interference of multiple clients
that would happen in OTRv2 without this feature. The original design of
otr4j would include a "default outgoing session", meaning that one could
send a message without specifying an instance and it would go to the
instance set as default outgoing instance. However, this means that the
host application must maintain correct awareness of whichever instance
is currently the "default outgoing instance" in order to properly
represent which instance is going to receive a message the user sends.
This is prone to error, causing misrepresentation of state in the UI,
because the host application's UI also needs to manage its current state.

Now that instance-tags are well established and we need to be able to
distinguish between instances, it becomes beneficial to remove this
implicit bit of state management. Instead, move all operations regarding
an outgoing session into its own interface `Instance`. The listener
interfaces already inform the host application of emergence of instances
and received messages include an instance-tag. Therefore, this new API
allows for making the outgoing instance explicit in your actions,
obviating the need for otr4j to do state management and instead
delegating state management to the host application.

This is the first introduction. Smaller changes may yet be necessary.